:recycle: turned images display on service details screen into a new component
